Removing remaining alerts (#1771)
Removed alerts that was tested on T2 and
should be on T1
